AU-1220: duplicates
